backend/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
dj-database-url
whitenoise
